Flytte validateDekorasjon og validateDekorasjonInnehavelse til models.py
